pyproject.toml
[tool.poetry]
name = "parking-lot"
version = "0.1.0"
description = ""
authors = ["Enias Oliveira <eniasoliveira27@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.2"
djangorestframework = "^3.12.4"
coverage = "^5.5"
ipdb = "^0.13.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
